Implement BaseLanguageModelParams interface that subclasses should extend
